#15 Move transfer result value object
