+++ bookman/domain/service.py
@@ -1,9 +1,8 @@
-from dataclasses import dataclass
-
 from django.db import transaction
 
 from bookman.domain.repository import BranchBookStockRepository
-from bookman.models import Book, Branch, BranchBookStock
+from bookman.domain.valueobject import BranchBookStockTransfer
+from bookman.models import Book, Branch
 
 
 class BranchBookStockTransferError(Exception):
@@ -22,28 +21,6 @@ class InsufficientStockError(BranchBookStockTransferError):
     """
     移動元支店の所蔵数が移動冊数に満たない場合の例外。
     """
-
-
-@dataclass(frozen=True)
-class BranchBookStockTransfer:
-    """
-    支店間移動後の所蔵状態。
-
-    Attributes:
-        book: 移動対象の書籍。
-        from_branch: 移動元支店。
-        to_branch: 移動先支店。
-        amount: 移動した冊数。
-        source_stock: 移動後の移動元支店別所蔵数。
-        destination_stock: 移動後の移動先支店別所蔵数。
-    """
-
-    book: Book
-    from_branch: Branch
-    to_branch: Branch
-    amount: int
-    source_stock: BranchBookStock
-    destination_stock: BranchBookStock
 
 
 class BranchBookStockTransferService:
